Splitting out generic helper functions
